icn_economics

Module marketplace

Structs

FulfillmentDetails

Details about how an item or service will be delivered.

InMemoryMarketplaceStore

In-memory marketplace store for testing and development.

LaborHoursConfig

Configuration for creating a labor hours marketplace offer

MarketplaceBid

Represents a bid to buy goods or services with tokens.

MarketplaceOffer

Represents an offer to sell goods or services for tokens.

MarketplaceTransaction

Represents a completed marketplace transaction.

OfferFilter

Filter criteria for searching marketplace offers.

PhysicalGoodConfig

Configuration for creating a physical good marketplace offer

ServiceConfig

Configuration for creating a service marketplace offer

Enums

BidStatus

Status of a marketplace bid.

FulfillmentMethod

Methods for fulfilling marketplace transactions.

ItemType

Types of items that can be traded in the marketplace.

OfferStatus

Status of a marketplace offer.

TransactionStatus

Status of a marketplace transaction.

Traits

MarketplaceStore

Trait for marketplace functionality.